requirements.txt
flask-Cors
Flask==2.1.0
gunicorn==20.1.0
opencv-python==4.6.0.66
google-cloud-storage==2.1.0
